adding trainer id verification

+++ src/modules/workout/services/DeleteWorkoutService.ts
@@ -10,11 +10,18 @@ class DeleteWorkoutService {
     private workoutRepository: IWorkoutRepository,
   ) {}
 
-  public async execute(id: string): Promise<void> {
+  public async execute(id: string, trainerId: string): Promise<void> {
     const workout = await this.workoutRepository.findById(id);
 
     if (!workout) {
       throw new AppError('Treino não encontrado', 404);
+    }
+
+    if (workout.trainerId !== trainerId) {
+      throw new AppError(
+        'Esse usuário não tem permissão para fazer alterações nesse treino',
+        403,
+      );
     }
 
     await this.workoutRepository.delete(id);
